requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
google-generativeai
googlesearch-python
pycountry
youtube_transcript_api
jina